Fix coverage by moving files between packages
