+++ python-frontend/src/test/java/org/sonar/plugins/python/IpynbNotebookParserTest.java
@@ -18,25 +18,37 @@ package org.sonar.plugins.python;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 import org.sonar.api.batch.fs.InputFile;
+import org.sonar.api.batch.fs.internal.TestInputFileBuilder;
 import org.sonar.api.internal.apachecommons.lang3.StringUtils;
-import org.sonar.plugins.python.api.IssueLocation;
-import org.sonar.plugins.python.api.PythonVisitorContext;
 import org.sonar.python.IPythonLocation;
-import org.sonar.python.TestPythonVisitorRunner;
-import org.sonar.python.checks.TrailingWhitespaceCheck;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.sonar.plugins.python.TestUtils.createInputFile;
-import static org.sonar.plugins.python.TestUtils.mapToColumnMappingList;
+import static org.sonar.plugins.python.NotebookTestUtils.mapToColumnMappingList;
 
 class IpynbNotebookParserTest {
   private final File baseDir = new File("src/test/resources/org/sonar/plugins/python").getAbsoluteFile();
 
+  public static PythonInputFile createInputFile(File baseDir, String name, InputFile.Status status, InputFile.Type type) {
+    try {
+      return new PythonInputFileImpl(TestInputFileBuilder.create("moduleKey", name)
+        .setModuleBaseDir(baseDir.toPath())
+        .setCharset(StandardCharsets.UTF_8)
+        .setStatus(status)
+        .setType(type)
+        .setLanguage("py")
+        .initMetadata(Files.readString(new File(baseDir, name).toPath()))
+        .build());
+    } catch (IOException e) {
+      throw new IllegalStateException("Cannot read " + name + " from base directory" + baseDir, e);
+    }
+  }
   @Test
   void testParseNotebook() throws IOException {
     var inputFile = createInputFile(baseDir, "notebook.ipynb", InputFile.Status.CHANGED, InputFile.Type.MAIN);
@@ -238,65 +250,4 @@ class IpynbNotebookParserTest {
     assertThat(result.contents()).isEmpty();
   }
 
-  @Test
-  void trailing_whitespace() throws IOException {
-    var inputFile = createInputFile(baseDir, "notebook_trailing_whitespace.ipynb", InputFile.Status.CHANGED, InputFile.Type.MAIN);
-    var result = IpynbNotebookParser.parseNotebook(inputFile).get();
-    var check = new TrailingWhitespaceCheck();
-    var context = new PythonVisitorContext(
-      TestPythonVisitorRunner.parseNotebookFile(result.locationMap(), result.contents()),
-      SonarQubePythonFile.create(result),
-      null,
-      "");
-    check.scanFile(context);
-
-    var issues = context.getIssues();
-    assertThat(issues).hasSize(3);
-
-    assertThat(issues.get(0).primaryLocation().startLine()).isEqualTo(14);
-    assertThat(issues.get(0).primaryLocation().endLine()).isEqualTo(14);
-    assertThat(issues.get(0).primaryLocation().startLineOffset()).isEqualTo(IssueLocation.UNDEFINED_OFFSET);
-    assertThat(issues.get(0).primaryLocation().endLineOffset()).isEqualTo(IssueLocation.UNDEFINED_OFFSET);
-
-    assertThat(issues.get(1).primaryLocation().startLine()).isEqualTo(17);
-    assertThat(issues.get(1).primaryLocation().endLine()).isEqualTo(17);
-    assertThat(issues.get(1).primaryLocation().startLineOffset()).isEqualTo(IssueLocation.UNDEFINED_OFFSET);
-    assertThat(issues.get(1).primaryLocation().endLineOffset()).isEqualTo(IssueLocation.UNDEFINED_OFFSET);
-
-    assertThat(issues.get(2).primaryLocation().startLine()).isEqualTo(20);
-    assertThat(issues.get(2).primaryLocation().endLine()).isEqualTo(20);
-    assertThat(issues.get(2).primaryLocation().startLineOffset()).isEqualTo(IssueLocation.UNDEFINED_OFFSET);
-    assertThat(issues.get(2).primaryLocation().endLineOffset()).isEqualTo(IssueLocation.UNDEFINED_OFFSET);
-  }
-
-  @Test
-  void trailing_whitespace_compressed() throws IOException {
-    var inputFile = createInputFile(baseDir, "notebook_trailing_whitespace_compressed.ipynb", InputFile.Status.CHANGED, InputFile.Type.MAIN);
-    var result = IpynbNotebookParser.parseNotebook(inputFile).get();
-    var check = new TrailingWhitespaceCheck();
-    var context = new PythonVisitorContext(
-      TestPythonVisitorRunner.parseNotebookFile(result.locationMap(), result.contents()),
-      SonarQubePythonFile.create(result),
-      null,
-      "");
-    check.scanFile(context);
-
-    var issues = context.getIssues();
-    assertThat(issues).hasSize(3);
-
-    assertThat(issues.get(0).primaryLocation().startLine()).isEqualTo(1);
-    assertThat(issues.get(0).primaryLocation().endLine()).isEqualTo(1);
-    assertThat(issues.get(0).primaryLocation().startLineOffset()).isEqualTo(142);
-    assertThat(issues.get(0).primaryLocation().endLineOffset()).isEqualTo(157); // Should be 154
-
-    assertThat(issues.get(1).primaryLocation().startLine()).isEqualTo(1);
-    assertThat(issues.get(1).primaryLocation().endLine()).isEqualTo(1);
-    assertThat(issues.get(1).primaryLocation().startLineOffset()).isEqualTo(199);
-    assertThat(issues.get(1).primaryLocation().endLineOffset()).isEqualTo(207); // Should be 204
-
-    assertThat(issues.get(2).primaryLocation().startLine()).isEqualTo(1);
-    assertThat(issues.get(2).primaryLocation().endLine()).isEqualTo(1);
-    assertThat(issues.get(2).primaryLocation().startLineOffset()).isEqualTo(249);
-    assertThat(issues.get(2).primaryLocation().endLineOffset()).isEqualTo(249); // Should be 255
-  }
 }
